Bump Ecto and Phoenix dependencies (#106)
* Bump Ecto and Phoenix

Add the needed dependencies as well as the explicit adapter that has to be passed
to the Ecto.Repo on repo.ex.
Resolves the Ecto.Multi problem which changed syntax and works differently in the
newest version.

* DELETE a user (attendee) correctly

If a user had badges then he couldn't be deleted because the if statement
would never be true. This is because Accounts.get_attendee! brings the attendee
associated with his badges, unlike user.attendee which brings only the
respective attendee without his badges.

* Make a function private

* Correct merge mistake

* Add some Badge tests

Add content-type to json response in an authentication error.

* Add tests for Delete Attendee

Change some peculiarities detected in tests.

* Add test for attendee with badges

* Change requested modification