Avatar links in Prizes and Redeemables (#198)
Since the badges are working fine on heroku, I tried to replicate the logic of the badge insertion task for both prizes and redeemables. I created the missing images for both structures, because the image links are being defined in avatar.ex, so if something is wrong it can't be there since that file is generic enough to handle any structure. However to test this hypothesis, this version needs to be on heroku. 😄

Probably solves: #193